happyruh_semantic_search/requirements.txt
# Web & API
fastapi==0.120.4
uvicorn[standard]==0.38.0
textblob==0.19.0
python-multipart==0.0.20

# Scraping
beautifulsoup4==4.12.3
requests==2.32.3

# AI Embedding + Semantic Search (PINNED for compatibility)
sentence-transformers==3.0.1
transformers==4.44.2
huggingface_hub==0.24.6
tokenizers==0.19.1
torch==2.2.2
numpy<2.0.0
qdrant-client==1.15.1

# PDF creation & parsing
reportlab==4.2.0
pymupdf==1.24.7

# DB & scheduling
psycopg2-binary==2.9.9
schedule==1.2.1

# Utilities
pydantic==2.7.1
python-dotenv==1.0.1
